Ensure nested statics have a HIR node to prevent various queries from ICEing
